requirements.txt
altair==5.5.0
anyio==4.7.0
apache-airflow-client==2.9.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
boto3==1.26.30
botocore==1.29.30
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==2.1.1
click==8.1.7
comm==0.2.2
debugpy==1.8.9
decorator==5.1.1
defusedxml==0.7.1
docker==6.0.0
executing==2.1.0
fastjsonschema==2.21.1
-e git+ssh://git@github.com/finout-io/finout-data-utils.git@f8468c51ea7bba56faa97dd44fc39a6b2619da50#egg=fincli&subdirectory=fincli
fqdn==1.5.1
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.37.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
hvac==0.10.6
idna==3.10
ipykernel==6.29.5
ipython==8.30.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.4
jmespath==1.0.1
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.3.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kubernetes==27.2.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.0.2
narwhals==1.34.0
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.3.1
notebook_shim==0.2.4
numpy==2.2.0
oauthlib==3.2.2
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
prettytable==3.11.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
protobuf==5.29.4
psutil==6.1.0
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.7
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydeck==0.9.1
Pygments==2.18.0
pylance==0.20.0
pyperclip==1.8.2
pyspark==3.5.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.2.0
pytz==2024.2
PyYAML==6.0.1
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.22.3
rsa==4.9
s3transfer==0.6.2
Send2Trash==1.8.3
setuptools==75.6.0
simple-term-menu==1.5.2
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
streamlit==1.44.1
tenacity==9.1.2
terminado==0.18.1
tinycss2==1.4.0
toml==0.10.2
tornado==6.4.2
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
urllib3==1.26.20
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.13
